* Fixing bug in FASTA reading   -- The bug was related with some sequence identifiers containing the `>` character in the middle. Our parser was not expecting that was considering that as the start of a new sequence.

+++ include/DistributedFastaData.h
@@ -33,7 +33,7 @@ class DistributedFastaData {
 public:
   ~DistributedFastaData();
 
-  DistributedFastaData(const char *file, ushort overlap, ushort k,
+  DistributedFastaData(const char *file, uint64_t overlap, ushort k,
                        const std::shared_ptr<ParallelOps> &parops,
                        const std::shared_ptr<TimePod> &tp, TraceUtils tu);
 
@@ -54,7 +54,7 @@ private:
   std::shared_ptr<TimePod> tp;
   TraceUtils tu;
   ushort k;
-  ushort overlap;
+  uint64_t overlap;
   FastaData *fd = nullptr;
 
   uint64_t l_seq_count;
