+++ include/ParallelFastaReader.hpp
@@ -12,7 +12,7 @@
 
 class ParallelFastaReader {
 public:
-  static void read_fasta(const char *file, ushort overlap, int rank,
+  static void read_fasta(const char *file, uint64_t overlap, int rank,
                   int world_size, char *&buff,
                   uint64_t &l_start, uint64_t &l_end);
 };
